+++ jscomp/frontend/external_ffi_types.mli
@@ -107,6 +107,8 @@ val inline_int64_primitive : int64 -> string list
 
 val inline_float_primitive : string -> string list
 
+val inline_bigint_primitive : string -> string list
+
 val ffi_bs : External_arg_spec.params -> return_wrapper -> external_spec -> t
 
 val ffi_bs_as_prims :