+++ jscomp/ml/predef.mli
@@ -32,6 +32,7 @@ val type_result: type_expr -> type_expr -> type_expr
 val type_dict: type_expr -> type_expr
 
 val type_int64: type_expr
+val type_bigint: type_expr
 val type_lazy_t: type_expr -> type_expr
 val type_extension_constructor:type_expr
 val type_floatarray:type_expr
@@ -51,10 +52,12 @@ val path_result: Path.t
 val path_dict: Path.t
 
 val path_int64: Path.t
+val path_bigint: Path.t
 val path_lazy_t: Path.t
 val path_extension_constructor: Path.t
 val path_floatarray: Path.t
 val path_promise: Path.t
+val path_uncurried: Path.t
 
 val path_match_failure: Path.t
 val path_assert_failure : Path.t