+++ jscomp/ml/untypeast.ml
@@ -114,7 +114,8 @@ let constant = function
   | Const_int i -> Pconst_integer (string_of_int i, None)
   | Const_int32 i -> Pconst_integer (Int32.to_string i, Some 'l')
   | Const_int64 i -> Pconst_integer (Int64.to_string i, Some 'L')
-  | Const_nativeint i -> Pconst_integer (Nativeint.to_string i, Some 'n')
+  | Const_bigint (sign, i) -> 
+    Pconst_integer (Bigint_utils.to_string sign i, Some 'n')
   | Const_float f -> Pconst_float (f,None)
 
 let attribute sub (s, p) = (map_loc sub s, p)