+++ jscomp/syntax/src/res_scanner.ml
@@ -274,14 +274,6 @@ let scanNumber scanner =
   (* suffix *)
   let suffix =
     match scanner.ch with
-    | 'n' ->
-      let msg =
-        "Unsupported number type (nativeint). Did you mean `" ^ literal ^ "`?"
-      in
-      let pos = position scanner in
-      scanner.err ~startPos:pos ~endPos:pos (Diagnostics.message msg);
-      next scanner;
-      Some 'n'
     | ('g' .. 'z' | 'G' .. 'Z') as ch ->
       next scanner;
       Some ch